Move unit test to separate file
